refactor : 디졸브 효과 FACharacter 로 승격
